+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+### Added
+ - Flatpak version of Steam is also detected with non-Flatpak installation of Protontricks
+
 ### Fixed
  - Fix detection of Steam library folders using non-standard capitalizations for `steamapps`
  - _Steam Linux Runtime - Sniper_ is no longer incorrectly reported as an unsupported runtime